Fix even more win64 warnings

+++ evutil_rand.c
@@ -121,6 +121,6 @@ void
 evutil_secure_rng_add_bytes(const char *buf, size_t n)
 {
 	arc4random_addrandom((unsigned char*)buf,
-	    n>(size_t)INT_MAX ? INT_MAX : n);
+	    n>(size_t)INT_MAX ? INT_MAX : (int)n);
 }
